new brand generator and share
